Création du GeofencingController
