+++ SampleApp/app/src/main/java/com/insiteo/sampleapp/MapFragment.java
@@ -61,7 +61,7 @@ import de.keyboardsurfer.android.widget.crouton.Crouton;
 import de.keyboardsurfer.android.widget.crouton.Style;
 
 public class MapFragment extends Fragment implements ISIMapListener, ISIRTOListener, OnClickListener,
-		ISIItineraryRendererListener, ISIItineraryRequestListener, ISILocationListener, ISIGeofenceListener, ActionBar.OnNavigationListener {
+		ISIItineraryRendererListener, ISIItineraryRequestListener, ISILocationListener, ActionBar.OnNavigationListener {
 
 	public final static String TAG = MapFragment.class.getSimpleName();
 
@@ -88,13 +88,18 @@ public class MapFragment extends Fragment implements ISIMapListener, ISIRTOListe
 	private boolean mCenterOnPosition = true;
 
 	// GEOFENCING
-	private ISGeofenceProvider mGeofenceProvider;
+	private GeofencingController geofencingController;
 	private View mGeofenceToastView = null;
 	private TextView mGeofenceToastText = null;
 
 	private int mLastLocMapID = -1;
-	
 
+
+	public MapFragment() {
+		MapViewController mapViewController = new MapViewController(this);
+		this.geofencingController = new GeofencingController(mapViewController);
+
+	}
 
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -108,10 +113,12 @@ public class MapFragment extends Fragment implements ISIMapListener, ISIRTOListe
 
 	@Override
 	public void onViewCreated(View view, Bundle savedInstanceState) {
+		mGeofenceToastView = LayoutInflater.from(getActivity()).inflate(R.layout.geofencing_toast, null);
+		mGeofenceToastText = (TextView)mGeofenceToastView.findViewById(R.id.geofence_text);
 		initializeMapService();
 		initializeLocationService();
 		initializeItineraryService();
-		initializeGeofencingService();
+		geofencingController.init();
 		super.onViewCreated(view, savedInstanceState);
 	}
 	
@@ -631,57 +638,6 @@ public class MapFragment extends Fragment implements ISIMapListener, ISIRTOListe
 	}
 
 	//******************************************************************************************************************
-	// GEOFENCING SERVICE 
-	// *****************************************************************************************************************
-
-	private void initializeGeofencingService(){
-		// geofencing
-		mGeofenceProvider = (ISGeofenceProvider) ISLocationProvider.getInstance().getModule(ISELocationModule.GEOFENCING);
-		mGeofenceProvider.setListener(this);
-		mGeofenceToastView = LayoutInflater.from(getActivity()).inflate(R.layout.geofencing_toast, null);
-		mGeofenceToastText = (TextView)mGeofenceToastView.findViewById(R.id.geofence_text);
-	}
-
-	/**
-	 * Called when geofencing module has new data available.
-	 * @param aEnteredZones list of zones that location has just entered  
-	 * @param aStayedZones list of zones where location has stayed for a certain amount of time
-	 * @param aLeftZones list of zones that location has just left
-	 */
-	@Override
-	public void onGeofenceUpdate(List<ISGeofenceArea> aEnteredZones, List<ISGeofenceArea> aStayedZones, List<ISGeofenceArea> aLeftZones) {
-		Log.d("Geofencing", "onGeofenceUpdate " + aEnteredZones.size() + " " + aStayedZones.size() + " " + aLeftZones.size());	
-
-		final List<ISGeofenceArea> zones = aEnteredZones;
-
-		for (ISGeofenceArea z : zones) {
-			final String extra1 = z.getExtra1();
-			if (extra1 != null) {
-				getActivity().runOnUiThread(new Runnable() {
-					@Override public void run() {
-						mGeofenceToastText.setText(extra1);
-						final Toast t = new Toast(getActivity());
-
-						t.setDuration(Toast.LENGTH_LONG);
-						Resources r = getResources();
-						float px = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 50, r.getDisplayMetrics());
-						t.setGravity(Gravity.TOP, 0, (int) px);
-						t.setView(mGeofenceToastView);
-						t.show();
-					}
-				});
-			}
-		}
-	}
-
-	/**
-	 * Called when geofencing data is reset. This happens when no location was received for a long duration, thus zone detection is not valid anymore.
-	 */
-	@Override
-	public void onGeofenceDataCleared() {
-	}
-
-	//******************************************************************************************************************
 	// LOCATION SERVICE
 	// *****************************************************************************************************************
 
@@ -853,4 +809,19 @@ public class MapFragment extends Fragment implements ISIMapListener, ISIRTOListe
 	}
 
 
+	public void showGeofenceToast(final String extra1) {
+		getActivity().runOnUiThread(new Runnable() {
+			@Override public void run() {
+				mGeofenceToastText.setText(extra1);
+				final Toast t = new Toast(getActivity());
+
+				t.setDuration(Toast.LENGTH_LONG);
+				Resources r = getResources();
+				float px = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 50, r.getDisplayMetrics());
+				t.setGravity(Gravity.TOP, 0, (int) px);
+				t.setView(mGeofenceToastView);
+				t.show();
+			}
+		});
+	}
 }
