+++ SampleApp/app/src/main/java/com/insiteo/sampleapp/MapViewController.java
@@ -1,0 +1,13 @@
+package com.insiteo.sampleapp;
+
+public class MapViewController {
+	private MapFragment mapFragment;
+
+	public MapViewController(MapFragment mapFragment) {
+		this.mapFragment = mapFragment;
+	}
+
+	public void showGeofenceToast(String extra1) {
+		mapFragment.showGeofenceToast(extra1);
+	}
+}
